+++ src/handle/tactic.ml
@@ -34,13 +34,14 @@ let add_axiom : Sig_state.t -> popt -> meta -> sym =
   let sym =
     wrn sym_pos "axiom %a: %a" uid name term !(m.meta_type);
     (* Temporary hack for axioms to have a declaration position in the order
-       they are created. *)
+       they are created, and strictly before the symbol. *)
+    (* FIXME: use sym_decl_pos instead ? *)
+    let id = {elt=name; pos=sym_pos} in
     let pos = shift Stdlib.(!admitted) sym_pos in
-    let id = Pos.make pos name in
     (* We ignore the new ss returned by Sig_state.add_symbol: axioms do not
        need to be in scope. *)
     snd (Sig_state.add_symbol ss
-           Public Defin Eager true id None !(m.meta_type) [] None)
+           Public Defin Eager true id pos !(m.meta_type) [] None)
   in
   (* Create the value which will be substituted for the metavariable. This
      value is [sym x0 ... xn] where [xi] are variables that will be
@@ -264,9 +265,7 @@ let get_config (ss:Sig_state.t) (pos:Pos.popt) : config =
 
 (** [p_term pos t] converts the term [t] into a p_term at position [pos]. *)
 let p_term (ss:Sig_state.t) (pos:popt): int StrMap.t -> term -> p_term =
-  let rec term idmap (t:term) :p_term =
-    (*if Logger.log_enabled() then log "p_term %a" Print.term t;*)
-    Pos.make pos (term_aux idmap t)
+  let rec term idmap (t:term) :p_term = Pos.make pos (term_aux idmap t)
   and params idmap x a =
     [Some(Pos.make pos (base_name x))],Some(term idmap a),false
   and term_aux idmap t :p_term_aux =
@@ -289,6 +288,7 @@ let p_term (ss:Sig_state.t) (pos:popt): int StrMap.t -> term -> p_term =
         let (x,b),idmap' = Print.safe_unbind idmap b in
         let id = Pos.make pos (base_name x) in
         P_LLet(id,[],Some(term idmap a),term idmap t,term idmap' b)
+    | Meta _ -> P_Wild
     | _ -> fatal pos "Unhandled term expression: %a." Print.term t
   in term
 
@@ -305,14 +305,21 @@ let p_ident_of_var (pos:popt) (t:term) :p_ident =
   | Vari v -> Pos.make pos (base_name v)
   | _ -> fatal pos "Not a variable of the proof context: %a." term t
 
+(* [pos_of_string s] assumes that [s] is a string literal and returns the
+   lexing position of the content of [s]. *)
+let pos_of_string: sym -> Lexing.position =
+  let f p = {p with start_offset=p.start_offset+1; start_col=p.start_col+1;
+                    end_offset=p.end_offset-1; end_col=p.end_col-1} in
+  fun s -> lexing_opt (Option.map f s.sym_pos)
+
 (** [p_term_of_string_term pos t] turns into a p_term a string literal term
     [t] that is part of a bigger term obtained by scoping and normalizing of a
     p_term at position [pos]. *)
 let p_term_of_string_term (pos:popt) (t:term): p_term =
   match t with
   | Symb s when String.is_string_literal s.sym_name ->
-    let p = lexing_opt (after s.sym_pos) in
-    Parsing.Parser.Lp.parse_term_string p (String.remove_quotes s.sym_name)
+    Parsing.Parser.Lp.parse_term_string (pos_of_string s)
+      (String.remove_quotes s.sym_name)
   | _ -> fatal pos "not a string literal"
 
 (** [p_rwpatt_of_string_term pos t] turns into a p_rwpatt option a string
@@ -324,9 +331,8 @@ let p_rwpatt_of_string_term (pos:popt) (t:term): p_rwpatt option =
   match t with
   | Symb s when String.is_string_literal s.sym_name ->
       let string = String.remove_quotes s.sym_name in
-      if string = "" then None
-      else let p = lexing_opt (after s.sym_pos) in
-           Some (Parsing.Parser.Lp.parse_rwpatt_string p string)
+      if string = "" then None else
+        Some (Parsing.Parser.Lp.parse_rwpatt_string (pos_of_string s) string)
   | _ -> fatal pos "not a string literal"
 
 (** [int_of_term pos t] returns the int contained in a string literal
@@ -342,94 +348,16 @@ let is_right (pos:popt) (t:term): bool =
   | "" | "right" -> true
   | _ -> fatal pos "invalid side literal"
 
-(** [p_tactic ss g env pos t] weak head normalizes [t] and convert the result
-    into a p_tactic. *)
-let p_tactic (ss:Sig_state.t) (g:goal) (env:Env.t) (pos:Pos.popt) (t:term)
-    : p_tactic =
-  let idmap = get_names g
-  and ctx = Env.to_ctxt env in
-  let c = get_config ss pos in
-  let p_term = p_term ss pos idmap in
-  let tac_eval t = Pos.make pos (P_tac_eval (p_term t)) in
-  let tac t =
-    let t = Eval.whnf ctx t in
-    if Logger.log_enabled() then log "reduces to: %a" term t;
-    match get_args t with
-    | Symb s, ts ->
-        begin
-          try
-            match Hashtbl.find c s.sym_name, ts with
-            | T_admit, _ -> P_tac_admit
-            | T_and, [t1;t2] -> P_tac_and(tac_eval t1, tac_eval t2)
-            | T_and, _ -> assert false
-            | T_all_hyps, [t] -> P_tac_all_hyps (p_term t)
-            | T_all_hyps, _ -> assert false
-            | T_apply, [_;t] -> P_tac_apply (p_term t)
-            | T_apply, _ -> assert false
-            | T_assume, [prefix;_;Abst (_, t)] ->
-              let v = new_var (string_of_term pos prefix) in
-              let n = Pos.make pos (uniq_name v) in
-              P_tac_and (Pos.make pos (P_tac_assume [Some n]),
-                         tac_eval (subst t (mk_Vari v)))
-            | T_assume, _ -> assert false
-            | T_assumption, [] -> P_tac_assumption
-            | T_assumption, _ -> assert false
-            | T_change, [_;t] -> P_tac_apply (p_term t)
-            | T_change, _ -> assert false
-            | T_fail, _ -> P_tac_fail
-            | T_first_hyp, [t] -> P_tac_first_hyp (p_term t)
-            | T_first_hyp, _ -> assert false
-            | T_focus, [t] -> P_tac_focus (string_of_term pos t)
-            | T_focus, _ -> assert false
-            | T_generalize, [_;t] -> P_tac_generalize(p_ident_of_var pos t)
-            | T_generalize, _ -> assert false
-            | T_have, [t1;t2] ->
-                let prf_sym = Builtin.get ss pos [] "P" in
-                let prf = p_term (mk_Symb prf_sym) in
-                let t2 = Pos.make pos (P_Appl(prf, p_term t2)) in
-                P_tac_have(Pos.make pos (string_of_term pos t1), t2)
-            | T_have, _ -> assert false
-            | T_induction, _ -> P_tac_induction
-            | T_orelse, [t1;t2] -> P_tac_orelse(tac_eval t1, tac_eval t2)
-            | T_orelse, _ -> assert false
-            | T_print, [t] ->
-              let arg =
-                match unfold t with
-                | Symb s ->
-                  let n = s.sym_name in
-                  if n = String.add_quotes "" then None
-                  else Some(Pos.make pos (s.sym_path, n))
-                | _ -> fatal pos "not a symbol or string literal: %a" term t
-              in
-              P_tac_query (Pos.make pos (P_query_print arg))
-            | T_print, _ -> assert false
-            | T_refine, [t] -> P_tac_refine(p_term_of_string_term pos t)
-            | T_refine, _ -> assert false
-            | T_reflexivity, _ -> P_tac_refl
-            | T_remove, [_;t] -> P_tac_remove [p_ident_of_var pos t]
-            | T_remove, _ -> assert false
-            | T_repeat, [t] -> P_tac_repeat(tac_eval t)
-            | T_repeat, _ -> assert false
-            | T_rewrite, [side;pat;_;t] ->
-                P_tac_rewrite(is_right pos side,
-                              p_rwpatt_of_string_term pos pat, p_term t)
-            | T_rewrite, _ -> assert false
-            | T_set, [t1;_;t2] ->
-                P_tac_set(Pos.make pos (string_of_term pos t1), p_term t2)
-            | T_set, _ -> assert false
-            | T_simplify, _ -> P_tac_simpl SimpAll
-            | T_simplify_beta, _ -> P_tac_simpl SimpBetaOnly
-            | T_solve, _ -> P_tac_solve
-            | T_symmetry, _ -> P_tac_sym
-            | T_try, [t] -> P_tac_try(tac_eval t)
-            | T_try, _ -> assert false
-            | T_why3, _ -> P_tac_why3 None
-          with Not_found ->
-            fatal pos "Unhandled tactic expression: %a." term t
-        end
-    | _ -> fatal pos "Unhandled tactic expression: %a." term t
-  in
-  Pos.make pos (tac t)
+(** [new_name prefix env] returns a string prefixed by [prefix] and not
+    occurring in [env]. *)
+let new_name (prefix:string) (env:Env.t): string =
+  if Env.mem prefix env then
+    let i = ref 0 in
+    let new_name() = prefix ^ string_of_int !i in
+    let s = ref (new_name()) in
+    while Env.mem !s env do incr i; s := new_name() done;
+    !s
+  else prefix
 
 (** [handle ss sym_pos priv ps tac] applies tactic [tac] in the proof state
    [ps] and returns the new proof state. *)
@@ -442,6 +370,105 @@ let handle (ss:Sig_state.t) (sym_pos:popt) (priv:bool)
       then Some new_ps
       else None
     with Fatal _ -> None
+
+  (* [p_tactic ss g env pos t] weak head normalizes [t] and converts the
+     result into a p_tactic. *)
+  and p_tactic (ps:proof_state) (g:goal) (env:Env.t) (pos:Pos.popt)
+    : term -> proof_state * p_tactic =
+    let idmap = get_names g
+    and ctx = Env.to_ctxt env in
+    let p_term = p_term ss pos idmap in
+    let mk = Pos.make pos in
+    let tac_eval t = mk(P_tac_eval(p_term t)) in
+    fun t ->
+      let t = Eval.whnf ctx t in
+      if Logger.log_enabled() then log "reduces to: %a" term t;
+      match get_args t with
+      | Symb s, ts ->
+        begin
+          try
+            (*FIXME: compute config only once in a proof*)
+            let c = get_config ss pos in
+            match Hashtbl.find c s.sym_name, ts with
+            | T_admit, _ -> ps, mk P_tac_admit
+            | T_and, [t1;t2] ->
+              let ps = handle ps (tac_eval t1) in ps, tac_eval t2
+            | T_and, _ -> assert false
+            | T_all_hyps, [t] -> ps, mk(P_tac_all_hyps(p_term t))
+            | T_all_hyps, _ -> assert false
+            | T_apply, [_;t] -> ps, mk(P_tac_apply(p_term t))
+            | T_apply, _ -> assert false
+            | T_assume, [prefix;_;Abst(_, t)] ->
+              begin
+                let n = new_name (string_of_term pos prefix) env in
+                let idopts = [Some(Pos.make pos n)] in
+                let new_ps = handle ps (mk (P_tac_assume idopts)) in
+                match new_ps.proof_goals with
+                | Typ{goal_hyps=(_,(v,_,_))::_;_}::_ ->
+                  new_ps, mk(P_tac_eval(p_term(subst t (mk_Vari v))))
+                | _ -> assert false
+              end
+            | T_assume, _ -> assert false
+            | T_assumption, [] -> ps, mk P_tac_assumption
+            | T_assumption, _ -> assert false
+            | T_change, [_;t] -> ps, mk(P_tac_apply (p_term t))
+            | T_change, _ -> assert false
+            | T_fail, _ -> ps, mk P_tac_fail
+            | T_first_hyp, [t] -> ps, mk(P_tac_first_hyp(p_term t))
+            | T_first_hyp, _ -> assert false
+            | T_focus, [t] -> ps, mk(P_tac_focus(string_of_term pos t))
+            | T_focus, _ -> assert false
+            | T_generalize, [_;t] ->
+              ps, mk(P_tac_generalize(p_ident_of_var pos t))
+            | T_generalize, _ -> assert false
+            | T_have, [t1;t2] ->
+                let prf_sym = Builtin.get ss pos [] "P" in
+                let prf = p_term (mk_Symb prf_sym) in
+                let t2 = Pos.make pos (P_Appl(prf, p_term t2)) in
+                ps, mk(P_tac_have(Pos.make pos (string_of_term pos t1), t2))
+            | T_have, _ -> assert false
+            | T_induction, _ -> ps, mk P_tac_induction
+            | T_orelse, [t1;t2] ->
+              ps, mk(P_tac_orelse(tac_eval t1, tac_eval t2))
+            | T_orelse, _ -> assert false
+            | T_print, [t] ->
+              let arg =
+                match unfold t with
+                | Symb s ->
+                  let n = s.sym_name in
+                  if n = String.add_quotes "" then None
+                  else Some(Pos.make pos (s.sym_path, n))
+                | _ -> fatal pos "not a symbol or string literal: %a" term t
+              in
+              ps, mk(P_tac_query (Pos.make pos (P_query_print arg)))
+            | T_print, _ -> assert false
+            | T_refine, [t] ->
+              ps, mk(P_tac_refine(p_term_of_string_term pos t))
+            | T_refine, _ -> assert false
+            | T_reflexivity, _ -> ps, mk P_tac_refl
+            | T_remove, [_;t] -> ps, mk(P_tac_remove [p_ident_of_var pos t])
+            | T_remove, _ -> assert false
+            | T_repeat, [t] -> ps, mk(P_tac_repeat(tac_eval t))
+            | T_repeat, _ -> assert false
+            | T_rewrite, [side;pat;_;t] ->
+              ps, mk(P_tac_rewrite(is_right pos side,
+                                   p_rwpatt_of_string_term pos pat, p_term t))
+            | T_rewrite, _ -> assert false
+            | T_set, [t1;_;t2] ->
+              let n = string_of_term pos t1 in
+              ps, mk(P_tac_set(Pos.make pos n, p_term t2))
+            | T_set, _ -> assert false
+            | T_simplify, _ -> ps, mk(P_tac_simpl SimpAll)
+            | T_simplify_beta, _ -> ps, mk(P_tac_simpl SimpBetaOnly)
+            | T_solve, _ -> ps, mk P_tac_solve
+            | T_symmetry, _ -> ps, mk P_tac_sym
+            | T_try, [t] -> ps, mk(P_tac_try(tac_eval t))
+            | T_try, _ -> assert false
+            | T_why3, _ -> ps, mk(P_tac_why3 None)
+          with Not_found ->
+            fatal pos "Unhandled tactic expression: %a." term t
+        end
+      | _ -> fatal pos "Unhandled tactic expression: %a." term t
 
   and handle ps ({elt;pos} as tac) =
   if Logger.log_enabled() then log "%a" Pretty.tactic tac;
@@ -516,7 +543,8 @@ let handle (ss:Sig_state.t) (sym_pos:popt) (priv:bool)
             let v = unfold (mk_Vari v) in
             let t = mk_Appl (mk_Appl (t, p), v) in
             if Logger.log_enabled () then log "ALL_HYPS on %a\n" term v;
-            try handle ps (p_tactic ss g env pos t) with Fatal _ -> ps
+            try let ps, t = p_tactic ps g env pos t in handle ps t
+            with Fatal _ -> ps
       in
       let ps' = List.fold_left try_assumption ps gt.goal_hyps in
       if ps' == ps then
@@ -575,7 +603,9 @@ let handle (ss:Sig_state.t) (sym_pos:popt) (priv:bool)
   | P_tac_first_hyp pt ->
     let t = scope pt in
     let f (_,(v,a,_)) =
-      progress ps (p_tactic ss g env pos (mk_Appl(mk_Appl(t,a),mk_Vari v))) in
+      let ps, t = p_tactic ps g env pos (mk_Appl(mk_Appl(t,a),mk_Vari v)) in
+      progress ps t
+    in
     begin match List.find_map f gt.goal_hyps with
     | None -> fatal pos "tactic [%a] fails on all assumptions" term t
     | Some new_ps -> new_ps
@@ -760,8 +790,10 @@ let handle (ss:Sig_state.t) (sym_pos:popt) (priv:bool)
           let term = term_in (Ctxt.names c) in
           fatal pt.pos "Cannot infer the type of [%a]" term t
       | Some(t,_) ->
-          if Unif.solve_noexn p then handle ps (p_tactic ss g env pos t)
-          else fatal pos "Cannot solve typing constraints for [%a]" term t
+        if Unif.solve_noexn p then
+          let ps, t = p_tactic ps g env pos t in handle ps t
+        else fatal pos "Cannot solve typing constraints for [%a]" term t
+
   in handle
 
 (** Representation of a tactic output. *)
@@ -786,30 +818,53 @@ let handle :
 
 (** [handle sym_pos priv r tac n] applies the tactic [tac] from the previous
    tactic output [r] and checks that the number of goals of the new proof
-   state is compatible with the number [n] of subproofs. *)
+   state is compatible with the number [n] of subproofs. When [tac] fails,
+   the proof state it was applied to is attached to the error (see
+   {!val:Proof.state_on_error}). *)
 let handle :
   Sig_state.t -> popt -> bool -> tac_output -> p_tactic -> int -> tac_output =
   fun ss sym_pos priv (ps, _) t nb_subproofs ->
-  let (ps', _) as a = handle ss sym_pos priv ps t in
+  (* Attach the proof state [t] was applied to, to any error escaping its
+     application. Errors raised inside tacticals like [try] or [orelse] are
+     caught below this point, so only failures actually reported to the user
+     are concerned. *)
+  let (ps', _) as a =
+    try handle ss sym_pos priv ps t
+    with Fatal(p, msg, desc)
+      when Stdlib.(!state_on_error) && ps.proof_goals <> [] ->
+        let state = error_state ps in
+        let desc = if desc = "" then state else desc ^ "\n" ^ state in
+        raise (Fatal(p, msg, desc))
+  in
   let nb_goals_before = List.length ps.proof_goals in
   let nb_goals_after = List.length ps'.proof_goals in
   let nb_newgoals = nb_goals_after - nb_goals_before in
+  (* [t] ran, but the number of subproofs given does not match the number of
+     subgoals it produced: report the proof state before and after its
+     application. *)
+  let mismatch : string -> 'a = fun reason ->
+    fatal t.pos ~err_desc:(error_state ~after:ps' ps) "%s" reason in
   if nb_newgoals <= 0 then
     if nb_subproofs = 0 then a
-    else fatal t.pos "A subproof is given but there is no subgoal."
+    else mismatch "A subproof is given but there is no subgoal."
   else if is_destructive t then
-    match nb_newgoals + 1 - nb_subproofs with
+    (match nb_newgoals + 1 - nb_subproofs with
     | 0 -> a
     | n when n > 0 ->
-      fatal t.pos "Missing subproofs (%d subproofs for %d subgoals):@.%a"
-        nb_subproofs (nb_newgoals + 1) goals ps'
+      mismatch (Printf.sprintf
+        "Missing subproofs (%d subproofs for %d subgoals)."
+        nb_subproofs (nb_newgoals + 1))
     | _ ->
-      fatal t.pos "Too many subproofs (%d subproofs for %d subgoals):@.%a"
-        nb_subproofs (nb_newgoals + 1) goals ps'
+      mismatch (Printf.sprintf
+        "Too many subproofs (%d subproofs for %d subgoals)."
+        nb_subproofs (nb_newgoals + 1)))
   else match nb_newgoals - nb_subproofs with
     | 0 -> a
     | n when n > 0 ->
-      fatal t.pos "Missing subproofs (%d subproofs for %d subgoals):@.%a"
-        nb_subproofs nb_newgoals goals ps'
-    | _ -> fatal t.pos "Too many subproofs (%d subproofs for %d subgoals)."
-             nb_subproofs nb_newgoals
+      mismatch (Printf.sprintf
+        "Missing subproofs (%d subproofs for %d subgoals)."
+        nb_subproofs nb_newgoals)
+    | _ ->
+      mismatch (Printf.sprintf
+        "Too many subproofs (%d subproofs for %d subgoals)."
+        nb_subproofs nb_newgoals)