+++ src/parsing/parser.ml
@@ -75,7 +75,7 @@ module Dk : PARSER with type lexbuf := Lexing.lexbuf = struct
     parse_lexbuf None entry lb
 
   let command =
-    let r = ref (Pos.none (P_open(false,[]))) in
+    let r = ref (Pos.none (P_open(None,false,[]))) in
     fun (lb:lexbuf): p_command ->
     Debug.(record_time Parsing
              (fun () -> r := DkParser.line DkLexer.token lb)); !r
@@ -114,15 +114,15 @@ sig
   end
 = struct
 
-  type zlexbuf = LpLexer.token ZipperTokenLexbuf.lexbuf
+  type zlexbuf = LpLexer.token OneLookaheadCellLexbuf.lexbuf
 
   let handle_error (icopt: in_channel option)
         (entry: zlexbuf -> 'a) (lb: zlexbuf): 'a option =
     try Some(entry lb)
     with
     | End_of_file -> Option.iter close_in icopt; None
-    | LpLexer.SyntaxError{pos=None; _} -> assert false
-    | LpLexer.SyntaxError{pos=Some pos; elt} ->
+    | LpLexer.SyntaxError(_,{pos=None; _}) -> assert false
+    | LpLexer.SyntaxError(_,{pos=Some pos; elt}) ->
         parser_fatal pos "Syntax error. %s" elt
 
   let parse_lexbuf' ~allow_rocq_syntax (icopt: in_channel option)
@@ -130,7 +130,7 @@ sig
     Stream.from
      (fun _ ->
        handle_error icopt entry
-        (ZipperTokenLexbuf.new_parser
+        (OneLookaheadCellLexbuf.new_parser
           ~pp:LpParser.pp_token
            ~lexer_:(fun () -> LpLexer.token ~allow_rocq_syntax lb)))
 
@@ -159,7 +159,6 @@ sig
     set_position lb lexpos;
     set_filename lb lexpos.pos_fname;
     Stream.next (parse_lexbuf' ~allow_rocq_syntax None entry lb)
-
 
   (* exported functions *)
   let parse_term_string =
@@ -192,7 +191,6 @@ let path_of_string : string -> Path.t = fun s ->
   with
       LpLexer.SyntaxError _ ->
       fatal_no_pos "Syntax error: \"%s\" is not a path." s
-
 
 (** [qident_of_string s] converts the string [s] into a qident. *)
 let qident_of_string : string -> Core.Term.qident = fun s ->